good progress

+++ examples/n_average.py
@@ -1,5 +1,4 @@
 import sd_mecha
-import torch
 sd_mecha.set_log_level()
 
 
@@ -16,5 +15,5 @@ recipe = models[0]
 for i, model in enumerate(models[1:], start=2):
     recipe = sd_mecha.weighted_sum(model, recipe, alpha=(i-1)/i)
 
-merger = sd_mecha.RecipeMerger(models_dir=r"E:\sd\models\Stable-diffusion", default_dtype=torch.float64)
+merger = sd_mecha.RecipeMerger(models_dir=r"E:\sd\models\Stable-diffusion")
 merger.merge_and_save(recipe, output="n_average")
